+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 * added following constants for [Code Execution](https://github.com/egodigital/vscode-powertools/wiki/Tools#code-execution):
   * `$code`: handles a value as string to display it in a webview with syntax highlight
+  * `$load`: loads data from an URI
   * `$tab`: opens a value as text in a new editor tab
+* fixes
 
 ## 0.29.0 (February 28th, 2019; Azure DevOps)